move updateLabel to Issue Tracker
