+++ backend/src/models/IssueTracker.ts
@@ -68,6 +68,40 @@ class Github extends IssueTracker {
             console.log(await response.json());
         })
     }
+
+    private addLabelToIssue(githubName: string, githubRepo: string, password: string, issueNumber, label: string) {
+        const link = `https://api.github.com/repos/${githubName}/${githubRepo}/issues/${issueNumber}/labels`;
+        const body = { labels: [label] };
+        const auth = `Basic ${Buffer.from(`${githubName}:${password}`, 'binary').toString('base64')}`;
+        fetch(link, {
+            method: 'post',
+            body: JSON.stringify(body),
+            headers: { Authorization: auth, 'Content-Type': 'application/json' }
+        })
+    }
+    
+    private removeLabelOfIssue(githubName: string, githubRepo: string, password: string, issueNumber, label: string) {
+        const link = `https://api.github.com/repos/${githubName}/${githubRepo}/issues/${issueNumber}/labels/${label}`;
+        const auth = `Basic ${Buffer.from(`${githubName}:${password}`, 'binary').toString('base64')}`;
+        fetch(link, {
+            method: 'post',
+            headers: { Authorization: auth, 'Content-Type': 'application/json' }
+        })
+    }
+    
+    updateLabel(testStatus: boolean, githubName: string, githubRepo: string, githubToken: string, issueNumber) {
+        let removeLabel;
+        let addedLabel;
+        if (testStatus) {
+            removeLabel = 'Seed-Test Test Fail :x:';
+            addedLabel = 'Seed-Test Test Success :white_check_mark:';
+        } else {
+            removeLabel = 'Seed-Test Test Success :white_check_mark:';
+            addedLabel = 'Seed-Test Test Fail :x:';
+        }
+        this.removeLabelOfIssue(githubName, githubRepo, githubToken, issueNumber, removeLabel);
+        this.addLabelToIssue(githubName, githubRepo, githubToken, issueNumber, addedLabel);
+    }
 }
 
 class NoTracker extends IssueTracker {
@@ -118,4 +152,6 @@ class Jira extends IssueTracker {
 export {
     IssueTrackerOption,
     IssueTracker,
+    Github,
+    Jira
 }
